Refactor filter component
