+++ src/components/BQGroupByRow.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
-import { BigQueryQueryNG, QueryWithDefaults } from 'types';
+import { BigQueryQueryNG, QueryWithDefaults } from '../types';
 import { useColumns } from '../utils/useColumns';
 import { useSqlChange } from '../utils/useSqlChange';
-import { SQLOrderByRow } from './visual-query-builder/SQLOrderByRow';
+import { SQLGroupByRow } from './visual-query-builder/SQLGroupByRow';
 
-type SQLOrderByRowProps = {
+interface BQGroupByRowProps {
   query: QueryWithDefaults;
   onQueryChange: (query: BigQueryQueryNG) => void;
-};
+}
 
-export function BQSQLOrderByRow({ query, onQueryChange }: SQLOrderByRowProps) {
+export function BQGroupByRow({ query, onQueryChange }: BQGroupByRowProps) {
   const columns = useColumns({ query, isOrderable: true });
   const { onSqlChange } = useSqlChange({ query, onQueryChange });
 
-  return <SQLOrderByRow sql={query.sql} onSqlChange={onSqlChange} columns={columns.value} />;
+  return <SQLGroupByRow columns={columns.value} sql={query.sql} onSqlChange={onSqlChange} />;
 }
