--- src/components/SQLBuilderWhereRow.scss
@@ -1,29 +1,0 @@
-@mixin flex($direction) {
-  display: flex;
-  /* autoprefixer: ignore next */
-  gap: 8px;
-  flex-direction: $direction;
-}
-
-.group--header {
-  @include flex(row);
-}
-
-.group-or-rule {
-  @include flex(column);
-  .rule {
-    flex-direction: row;
-  }
-}
-
-.rule--body {
-  @include flex(row);
-}
-
-.group--children {
-  @include flex(column);
-}
-
-.group--conjunctions:empty {
-  display: none;
-}
